After discussion with the authors, this is better
The difference appears with alpha.bst. Entering the author name in the
bst as {El Housni} makes it an atom as far as bibtex is concerned, and
this changes the notion of what the "initial" of this atom is. The
author wants the alpha.bst key rendered as EH. To achieve this, bibtex
needs to recognize "EH" as the initial, and this is possible only is
bibtex recognizes a multi-token "last" name. This can only be achieved
with this format. (see also the very last sentence of section 4 in
http://mirrors.ctan.org/biblio/bibtex/contrib/doc/btxdoc.pdf)